requirements.txt
Flask>=2.0
google-generativeai>=0.4.0 # Or the latest version
python-dotenv>=0.19.0     # Good practice for managing API keys